requirements.txt
pandas>=1.5.0
numpy>=1.21.0
nltk>=3.8.1
spacy>=3.7.2
beautifulsoup4>=4.12.0
requests>=2.31.0
newsapi-python>=0.2.7
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0
textblob>=0.17.1
scikit-learn>=1.0.0
plotly>=5.13.0
dash>=2.9.0 